+++ src/ft_printf/ft_snprintf/sprint_ptr.c
@@ -6,7 +6,7 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 19:27:12 by ldulling          #+#    #+#             */
-/*   Updated: 2025/01/29 09:49:27 by ldulling         ###   ########.fr       */
+/*   Updated: 2025/01/30 07:44:54 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 #include "libft.h"
 #include <stddef.h>
 
-static void	sprint_nullptr(t_sformat *f);
-static int	ptrlen(size_t ptr);
-static int	fullptrlen(int len_ptr, t_sformat *f);
-static void	sputhex(size_t ptr, t_sformat *f);
+static void			sprint_nullptr(t_sformat *f);
+static unsigned int	ptrlen(size_t ptr);
+static unsigned int	fullptrlen(unsigned int len_ptr, t_sformat *f);
+static void			sputhex(size_t ptr, t_sformat *f);
 
 void	sprint_ptr(size_t ptr, t_sformat *f)
 {
-	int	len_ptr;
-	int	len_full;
+	unsigned int	len_ptr;
+	unsigned int	len_full;
 
 	if (!ptr)
 		return (sprint_nullptr(f));
@@ -36,7 +36,7 @@ void	sprint_ptr(size_t ptr, t_sformat *f)
 	else if (f->space)
 		f->sprinted += ft_sputnchar(&f->str[f->sprinted], ' ', max_size(f, 1));
 	f->sprinted += ft_sputnstr(&f->str[f->sprinted], "0x", max_size(f, 2));
-	if (f->precision > len_ptr)
+	if (f->precision > (int)len_ptr)
 		f->sprinted += ft_sputnchar(&f->str[f->sprinted], '0',
 				max_size(f, f->precision - len_ptr));
 	else if (f->zero && !f->minus && f->precision < 0 && f->width > len_full)
@@ -50,7 +50,7 @@ void	sprint_ptr(size_t ptr, t_sformat *f)
 
 static void	sprint_nullptr(t_sformat *f)
 {
-	int	len;
+	unsigned int	len;
 
 	len = ft_strlen(NULL_PRINTOUT_PTR);
 	if (!f->minus && f->width > len)
@@ -63,9 +63,9 @@ static void	sprint_nullptr(t_sformat *f)
 				max_size(f, f->width - len));
 }
 
-static int	ptrlen(size_t ptr)
+static unsigned int	ptrlen(size_t ptr)
 {
-	int	len_ptr;
+	unsigned int	len_ptr;
 
 	len_ptr = 0;
 	while (ptr)
@@ -76,12 +76,12 @@ static int	ptrlen(size_t ptr)
 	return (len_ptr);
 }
 
-static int	fullptrlen(int len_ptr, t_sformat *f)
+static unsigned int	fullptrlen(unsigned int len_ptr, t_sformat *f)
 {
-	int	len_full;
+	unsigned int	len_full;
 
 	len_full = len_ptr;
-	if (len_full < f->precision)
+	if ((int)len_full < f->precision)
 		len_full = f->precision;
 	if (f->plus || f->space)
 		len_full++;
