+++ src/ft_printf/ft_snprintf/sprint_str.c
@@ -6,36 +6,37 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 19:27:17 by ldulling          #+#    #+#             */
-/*   Updated: 2025/01/29 09:47:38 by ldulling         ###   ########.fr       */
+/*   Updated: 2025/01/29 16:44:54 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft.h"
+#include <stddef.h>
 
-static void	sprint(const char *str, int len, t_sformat *f);
+static void	sprint(const char *str, size_t len, t_sformat *f);
 
 void	sprint_str(const char *str, t_sformat *f)
 {
-	int	len;
+	size_t	len;
 
 	if (!str)
 	{
 		len = ft_strlen(NULL_PRINTOUT_STR);
-		if (f->precision >= 0 && len > f->precision)
+		if (f->precision >= 0 && len > (size_t)f->precision)
 			len = 0;
 		sprint(NULL_PRINTOUT_STR, len, f);
 	}
 	else
 	{
 		len = ft_strlen(str);
-		if (f->precision >= 0 && len > f->precision)
+		if (f->precision >= 0 && len > (size_t)f->precision)
 			len = f->precision;
 		sprint(str, len, f);
 	}
 }
 
-static void	sprint(const char *str, int len, t_sformat *f)
+static void	sprint(const char *str, size_t len, t_sformat *f)
 {
 	if (!f->minus && f->width > len)
 		f->sprinted += ft_sputnchar(&f->str[f->sprinted], ' ',
